pyproject.toml
[project]
name = "throw-iniciative"
version = "0.1.0"
description = "Sistema de RPG com iniciativa d20 em tempo real"
authors = [
    {name = "Yahto"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2)",
    "djangorestframework (>=3.15)",
    "django-cors-headers (>=4.4)",
    "python-decouple (>=3.8)",
    "channels (>=4.1)",
    "psycopg2-binary (>=2.9)",
    "numpy (>=2.0)",
    "pandas (>=2.2)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
